// src/types/mock-data.type.ts
export type MockDataType = {
  titles: string[];
  descriptions: string[];
  dates: string[];
  cities: string[];
  previews: string[];
  photoes: string[];
  offerTypes: string[];
  advantages: string[];
  names: string[];
  emails: string[];
  avatars: string[];
  passwords: string[];
  types: ('pro' | 'base')[];
  latitude: number[];
  longitude: number[];
}